% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_aovlist.R
\name{anova.aovlist}
\alias{anova.aovlist}
\title{ANOVA for an aovlist object}
\usage{
\method{anova}{aovlist}(object, ...)
}
\arguments{
\item{object}{Model of class \code{aovlist}}

\item{...}{Further arguments to \code{stats:::summary.aovlist}}
}
\description{
ANOVA for an aovlist object
}
