% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_aovlist.R
\name{fitted.aovlist}
\alias{fitted.aovlist}
\title{Fitted values for an aovlist object}
\usage{
\method{fitted}{aovlist}(object, ...)
}
\arguments{
\item{object}{Model of class \code{aovlist}}

\item{...}{Further arguments to \code{dae::fitted.aovlist}}
}
\description{
Fitted values for an aovlist object
}
