% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_aovlist.R
\name{plot.aovlist}
\alias{plot.aovlist}
\title{Diagnostic plots for an aovlist object}
\usage{
\method{plot}{aovlist}(x, ...)
}
\arguments{
\item{x}{Model of class \code{aovlist}}

\item{...}{Further arguments to \code{plot}}
}
\description{
Diagnostic plots for an aovlist object
}
