% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralANCOVA.R
\name{PosthocTests}
\alias{PosthocTests}
\title{Callback function for Posthoc tests menu (activates on general ANCOVA model)}
\usage{
PosthocTests()
}
\description{
Callback function for Posthoc tests menu (activates on general ANCOVA model)
}
