% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_aovlist.R
\name{residuals.aovlist}
\alias{residuals.aovlist}
\title{Residual values for an aovlist object}
\usage{
\method{residuals}{aovlist}(object, ...)
}
\arguments{
\item{object}{Model of class \code{aovlist}}

\item{...}{Further arguments to \code{dae::residuals.aovlist}}
}
\description{
Residual values for an aovlist object
}
