% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralANCOVA.R
\name{TableOfEffects}
\alias{TableOfEffects}
\title{Callback function for Tble of Effects menu (activates on general ANCOVA model)}
\usage{
TableOfEffects()
}
\description{
Callback function for Tble of Effects menu (activates on general ANCOVA model)
}
